+++ test/maskIsolatedTest.cpp
@@ -3,8 +3,8 @@
 #include "catch.hpp"
 
 struct Event {
-    uint_fast16_t x;
-    uint_fast16_t y;
+    uint64_t x;
+    uint64_t y;
     uint_fast64_t timestamp;
 };
