+++ test/mirrorXTest.cpp
@@ -3,7 +3,7 @@
 #include "catch.hpp"
 
 struct Event {
-    uint_fast16_t x;
+    uint64_t x;
 };
 
 TEST_CASE("Invert the x coordinate", "[MirrorX]") {
